Clean up all compiler warnings (86 -> 0)
Our own code is now warning-free; third-party noise is scoped away:

- TextureEffectKaleidoscopeBase.cpp: add f suffixes to the float member
  initialisers / assignments that took double literals (52x C4305).
- printf format fixes (C4477): Configuration.cpp uses qPrintable() for a QString
  %s; filterShader.cpp casts m_imageList.size() for %d.
- size_t -> int/unsigned narrowing casts (C4267) in EffectShader, glwidget, main;
  signed/unsigned compare fixed by making maxStringlength int (C4018).
- EffectShader.cpp: m_uniforms.empty() (result discarded, a no-op - C4858) was a
  latent bug; changed to the intended clear().
- shader_setup.cpp: strdup -> _strdup (C4996).
- GLee.c (vendored): per-file WarningLevel = TurnOffAllWarnings (4x C4267).
- Scoped DisableSpecificWarnings 4996 on the x64 configs to silence Qt 6.11's
  own QGuiApplication::compressEvent Qt-7 deprecation notice (we don't use it;
  CRT deprecations are already off via _CRT_SECURE_NO_WARNINGS).

Rebuild is now 0 warnings, 0 errors; smoke-tested, runs unchanged.

+++ filterShader.cpp
@@ -162,7 +162,7 @@ void FilterShader::start( int width, int height )
 	traverse( m_imageDirectory, m_imageList );
 	m_imageListIterator = m_imageList.begin();
 
-	printf( "Nr of images: %d\n", m_imageList.size() );
+	printf( "Nr of images: %d\n", (int) m_imageList.size() );
 	if( m_imageList.isEmpty() )
 		fprintf( stderr, "WARNING: image directory '%s' missing or empty - "
 		                 "using a procedural fallback texture.\n",
@@ -344,7 +344,7 @@ FilterShader::FilterShader(int width, int height, const QString &directory)
 	traverse( directory, m_imageList );
 	m_imageListIterator = m_imageList.begin();
 
-	printf( "Nr of images: %d\n", m_imageList.size() );
+	printf( "Nr of images: %d\n", (int) m_imageList.size() );
 
 	qsrand(0);  // no-op: QRandomGenerator is auto-seeded
     unsigned int start = qrand() % (m_imageList.size() + 1);
